requirements.txt
pyteal==0.24.1
Beaker==1.12.1
fastapi==0.95.1
python-dotenv==1.0.0
Flask==2.2.3
PyJWT==2.6.0
backend==0.2.4.1
Werkzeug==2.2.3
SQLAlchemy==2.0.10
PyMySQL==1.0.3
bcrypt~=4.0.1